chore: Popper officially in the Popup module
