+++ 4.0-MIGRATION-GUIDE.md
@@ -333,8 +333,8 @@ PRs:
 
 `Popper` was changed to a Functional Component with a forwarded ref. If you passed a `ref` object to
 `Popper` before, it will now point to the element rather than the `Popper` instance. Popper was
-moved to the `@workday/canvas-kit-react-popup module. It will remain in common to prevent
-unnecessary breaking changes.
+moved to the `@workday/canvas-kit-react-popup module. This change aligns with the concept that Popup
+is a type of UI behavior. Popups can be built on top of the popup system in the Popup module.
 
 PRs:
