+++ modules/common/react/index.ts
@@ -1,7 +1,6 @@
 export * from './lib/parts';
 export * from './lib/types';
 export * from './lib/genericStyles';
-export * from './lib/Popper';
 export * from './lib/theming';
 export * from './lib/styles';
 export * from './lib/utils';
