+++ modules/common/react/lib/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './colorUtils';
-export * from './getTransformFromPlacement';
 export * from './getTranslateFromOrigin';
 export * from './makeMq';
