--- modules/common/react/stories/stories.tsx
@@ -1,59 +1,0 @@
-/// <reference path="../../../../typings.d.ts" />
-import * as React from 'react';
-import {storiesOf} from '@storybook/react';
-import withReadme from 'storybook-readme/with-readme';
-
-import {Button} from '../../../button/react';
-import {Popup} from '../../../popup/react';
-import {Popper} from '../index';
-import README from '../README.md';
-
-interface PopperPopupState {
-  open: boolean;
-}
-
-class PopperPopup extends React.Component<{}, PopperPopupState> {
-  private buttonRef: React.RefObject<HTMLButtonElement> = React.createRef();
-
-  public state: PopperPopupState = {
-    open: false,
-  };
-
-  public render() {
-    const {open} = this.state;
-
-    return (
-      <div style={{display: 'flex', justifyContent: 'center'}}>
-        <Button
-          buttonRef={this.buttonRef}
-          variant={Button.Variant.Primary}
-          onClick={this.handleClick}
-        >
-          Toggle Popup
-        </Button>
-        <Popper placement={'bottom'} open={open} anchorElement={this.buttonRef.current!}>
-          <Popup width={400} heading={'Popper Example'} handleClose={this.handleClose}>
-            <h3>Welcome to your popup positioned by Popper!</h3>
-          </Popup>
-        </Popper>
-      </div>
-    );
-  }
-
-  private handleClose = () => {
-    this.setState({open: false});
-  };
-
-  private handleClick = () => {
-    this.setState({open: !this.state.open});
-  };
-}
-
-storiesOf('Components|Popups/Popper/React', module)
-  .addParameters({component: Popper})
-  .addDecorator(withReadme(README))
-  .add('Popper', () => (
-    <div className="story">
-      <PopperPopup />
-    </div>
-  ));
