+++ modules/popup/react/stories/stories.tsx
@@ -7,91 +7,59 @@ import {Button, DeleteButton} from '@workday/canvas-kit-react-button';
 import {Popper, Popup} from '@workday/canvas-kit-react-popup';
 
 import README from '../README.md';
+import {useOutsideClick} from '../lib/useOutsideClick';
+import {useEscapeKey} from '../lib/useEscapeKey';
 
-interface PopupWrapperState {
-  open: boolean;
-  anchorEl: HTMLElement | null;
-}
+export default {
+  title: 'Components|Popups/Popup/React',
+  component: Popup,
+  decorators: [withReadme(README)],
+};
 
-class PopupWrapper extends React.Component<{}, PopupWrapperState> {
-  state = {
-    open: false,
-    anchorEl: null,
+export const Default = () => {
+  const [open, setOpen] = React.useState(false);
+  const [anchorEl, setAnchorEl] = React.useState<HTMLElement>(null);
+  const popupRef = React.useRef<HTMLDivElement>(null);
+
+  const onClose = () => setOpen(false);
+  const onButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    setOpen(true);
+    setAnchorEl(event.currentTarget);
   };
 
-  public render() {
-    const {anchorEl, open} = this.state;
-    return (
-      <div style={{display: 'flex', justifyContent: 'center'}}>
-        <DeleteButton onClick={this.handleClick}>Delete Item</DeleteButton>
-        <Popper placement={'bottom'} open={open} anchorElement={anchorEl}>
-          <Popup
-            width={400}
-            heading={'Delete Item'}
-            padding={Popup.Padding.s}
-            handleClose={this.handleClose}
-          >
-            <div style={{marginBottom: '24px'}}>
-              Are you sure you'd like to delete the item titled 'My Item'?
-            </div>
+  useOutsideClick(popupRef, onClose);
+  useEscapeKey(popupRef, onClose);
 
-            <DeleteButton style={{marginRight: '16px'}} onClick={this.handleSubmit}>
-              Delete
-            </DeleteButton>
-            <Button onClick={this.handleSubmit}>Cancel</Button>
-          </Popup>
-        </Popper>
-      </div>
-    );
-  }
+  return (
+    <div style={{display: 'flex', justifyContent: 'center'}}>
+      <DeleteButton onClick={onButtonClick}>Delete Item</DeleteButton>
+      <Popper placement={'bottom'} open={open} anchorElement={anchorEl} ref={popupRef}>
+        <Popup width={400} heading={'Delete Item'} padding={Popup.Padding.s} handleClose={onClose}>
+          <div style={{marginBottom: '24px'}}>
+            Are you sure you'd like to delete the item titled 'My Item'?
+          </div>
 
-  private handleClose = () => {
-    this.setState({
-      open: false,
-    });
-  };
-
-  private handleSubmit = () => {
-    this.setState({
-      open: false,
-    });
-  };
-
-  private handleClick = (e: React.MouseEvent<HTMLButtonElement>) => {
-    const {currentTarget} = e;
-    this.setState({
-      anchorEl: currentTarget,
-      open: !this.state.open,
-    });
-  };
-}
-
-storiesOf('Components|Popups/Popup/React', module)
-  .addParameters({component: Popup})
-  .addDecorator(withReadme(README))
-  .add('Default', () => (
-    <div className="story">
-      <PopupWrapper />
+          <DeleteButton style={{marginRight: '16px'}} onClick={onClose}>
+            Delete
+          </DeleteButton>
+          <Button onClick={onClose}>Cancel</Button>
+        </Popup>
+      </Popper>
     </div>
-  ))
-  .add(
-    'Open',
-    () => (
-      <Popup width={400} heading={'Delete Item'} padding={Popup.Padding.s} handleClose={() => null}>
-        <div style={{marginBottom: '24px'}}>
-          Are you sure you'd like to delete the item titled 'My Item'?
-        </div>
-
-        <DeleteButton style={{marginRight: '16px'}} onClick={() => null}>
-          Delete
-        </DeleteButton>
-        <Button onClick={() => null}>Cancel</Button>
-      </Popup>
-    ),
-    {
-      chromatic: {
-        viewports: [320],
-        pauseAnimationAtEnd: true,
-      },
-    }
   );
+};
+
+export const Open = () => {
+  return (
+    <Popup width={400} heading={'Delete Item'} padding={Popup.Padding.s} handleClose={() => null}>
+      <div style={{marginBottom: '24px'}}>
+        Are you sure you'd like to delete the item titled 'My Item'?
+      </div>
+
+      <DeleteButton style={{marginRight: '16px'}} onClick={() => null}>
+        Delete
+      </DeleteButton>
+      <Button onClick={() => null}>Cancel</Button>
+    </Popup>
+  );
+};
